+++ fabric/src/main/java/dev/qixils/crowdcontrol/plugin/fabric/mixin/ServerGamePacketListenerImplMixin.java
@@ -1,0 +1,23 @@
+package dev.qixils.crowdcontrol.plugin.fabric.mixin;
+
+import dev.qixils.crowdcontrol.plugin.fabric.commands.LootboxCommand;
+import net.minecraft.network.protocol.game.ServerboundContainerClosePacket;
+import net.minecraft.server.level.ServerPlayer;
+import net.minecraft.server.network.ServerGamePacketListenerImpl;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(ServerGamePacketListenerImpl.class)
+public class ServerGamePacketListenerImplMixin {
+
+	@Shadow
+	public ServerPlayer player;
+
+	@Inject(method = "handleContainerClose", at = @At("HEAD"))
+	public void onHandleContainerClose(ServerboundContainerClosePacket serverboundContainerClosePacket, CallbackInfo ci) {
+		LootboxCommand.onInventoryClose(player);
+	}
+}